extract submit handler into separate hook
